+++ .ci/docs.sh
@@ -6,7 +6,7 @@ if [[ $1 == "install" ]]; then
   conda install -q pandoc
 elif [[ $1 == "script" ]]; then
   git clone -b gh-pages-release https://github.com/nengo/nengo-dl.git ../nengo-dl-docs
-  RELEASES=$(find ../nengo-dl-docs -maxdepth 1 -type d -name "v[0-9].*" -printf "%f;")
+  RELEASES=$(find ../nengo-dl-docs -maxdepth 1 -type d -name "v[0-9].*" -printf "%f,")
 
   if [[ "$TRAVIS_BRANCH" == "$TRAVIS_TAG" ]]; then
     RELEASES="$RELEASES$TRAVIS_TAG"